luopeng_add the test code in C++_3

+++ C++_3/simon-speck.h
@@ -29,7 +29,8 @@
 typedef  unsigned char u8;
 typedef  unsigned short u16;
 typedef  unsigned int u32;
-typedef  uint64_t u64;
+typedef  unsigned long long u64;
+// typedef  uint64_t u64;
 
 /*
  * rotate shift left x by n bits
